+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Breaking]: Introduced the `GoogleApiConfig` class to manage the Google Places API Configuration.
 * [Breaking]: Renamed `onPlaceDetailsWithCoordinatesReceived` to `onPredictionWithCoordinatesReceived`.
+* [Breaking]: Bumped up SDK version to `3.7.2`.
 * [Feat]: Add `locationBias` and `locationRestriction` parameters to the api config(see more info [in the Google docs](https://developers.google.com/maps/documentation/places/web-service/place-autocomplete#location-bias-restriction))
 * [Feat]: Add `includedPrimaryTypes` parameter to the api config (see more info [in the Google docs](https://developers.google.com/maps/documentation/places/web-service/place-autocomplete#includedPrimaryTypes))